Bubble game by elaigwuresource
Bubble game created by elaigwuresource team. A game designed for the player to know and learn the gimmicks behind navigating to surpass enemies, as the the fish swim between several enemies and pop up bubbles to earn scores, this a level one design. Navigation and bubble popping to earn scores is a major plus for the players.Just in brief description